refactor(studio): inject runtime store dependencies
