+++ scripts/qa/check_file_sizes.py
@@ -54,9 +54,7 @@ LEGACY_LIMITS: Final = {
     "src/contexts/ai/infrastructure/providers/dashscope_text_generation_provider.py": 633,
     "tests/contexts/ai/infrastructure/test_text_generation_providers.py": 570,
     "src/contexts/studio/application/ports/studio_repository.py": 543,
-    "src/contexts/studio/interface/http/router.py": 533,
     "src/shared/infrastructure/config/settings.py": 515,
-    "src/contexts/studio/application/services/facade.py": 429,
     "src/contexts/studio/application/services/ai_service.py": 377,
     "src/contexts/studio/application/service_common.py": 354,
     "tests/contexts/studio/test_services.py": 316,
